+++ internal/Panels/useFocusOnTransition.js
